refactor!: split lerna and single-repository configurations further
BREAKING CHANGES: Release notes will be sparse because this project is
not officially on its last legs. I want to start moving towards projen,
but have not the bandwidth to initiate that project at the moment,
so I'm making the minimal (and likely undocumented) set of changes
to keep this project sputtering along on fumes for the time being.

Please let me know if you're interested in taking over maintenance
of this project!

- rename prettier* npm scripts to format*
- remove eslint and prettier packages from the generated lerna manifests

Closes #216, closes #209, closes #205